fix(animation): address review findings on exit timing and geometry
Several surfaces built their closing animation from the enter token while
their presence unmounted on the exit token, so the visual animation was
still running when its content disappeared. The accordion did this for
both its layout and chevron animations, and its presence window also
skipped the unmount grace; the sidebar submenu sprang back on collapse
where its own chevron correctly used the exit token.

The select dropdown slid the opposite way from every other flyout:
Anchor::Top* means the menu sits below the trigger, which popover and
hover card give -1.0, while select matched on Anchor::Bottom*.

Command palette rows gained a 1px border only when selected inside a
fixed-height row, shrinking the content box and shifting the label as
selection moved. The border is now reserved on every row and only its
color changes.

exit_duration added two u16 values before widening. exit_duration_ms is
deserialized from theme JSON, so a large value would wrap to a near-zero
exit window and unmount on the first frame.

Both progress controls rebuilt the standard curve inline from
enter_duration_ms and standard_easing rather than calling
standard_animation, so retuning that curve would have skipped them.

Also fixes the floating sidebar story, where max() against the enter
duration made the collapse branch a no-op, and a doc comment that counted
five easing curves where there are four.

Found by CodeRabbit review on #16.

+++ crates/ui/src/animation.rs
@@ -163,7 +163,10 @@ const EXIT_UNMOUNT_GRACE_MS: u16 = 33;
 /// element mounted so [`exit_animation`] plays to completion. Slightly longer
 /// than the animation itself; see [`EXIT_UNMOUNT_GRACE_MS`].
 pub fn exit_duration(motion: &ThemeMotion) -> Duration {
-    Duration::from_millis(u64::from(motion.exit_duration_ms + EXIT_UNMOUNT_GRACE_MS))
+    // Widen before adding: exit_duration_ms is deserialized from theme JSON, so
+    // a value near u16::MAX would otherwise wrap to a near-zero exit window and
+    // unmount the element on its first frame.
+    Duration::from_millis(u64::from(motion.exit_duration_ms) + u64::from(EXIT_UNMOUNT_GRACE_MS))
 }
 
 /// The one spring, for transform-only reveal motion. Settles within the `enter`
